backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
landingai-ade
pydantic
anthropic
